removeTajimaD
